`createFlatProxy`
